Update start.yml
This start configuration is 791 tokens (about the size of the original one I had), but contains much more information about its capabilities. It also seems to be more aligned with the vision of Valyxa. Namely, digesting a user's goal, splitting it into a project with routines, notes, etc., and creating them.